# contracts/openapi/v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SLOPwise legacy API",
    "version": "1.0.0.2026-09-13",
    "description": "Canonical contract of the current Next.js /api/v1 API. Monetary integer fields use the application's current minor-unit convention."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin Next.js server"
    }
  ],
  "security": [
    {
      "sessionCookie": []
    }
  ],
...
  "paths": {
...
    "/api/v1/users/register": {
      "post": {
        "operationId": "registerUserV1",
        "tags": ["Users"],
        "summary": "Register a credentials user",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterUserRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registered public user",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RegisterUserResponse" }
              }
            }
          },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/UnprocessableEntity" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "authjs.session-token",
        "description": "Auth.js same-origin session cookie. Secure deployments may use the __Secure- prefixed cookie name."
      }
    },
...
    "responses": {
...
      "Conflict": {
        "description": "Domain conflict, duplicate registration, or exhausted Serializable retries",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/LegacyErrorEnvelope" }
          }
        }
      },
      "UnprocessableEntity": {
        "description": "Zod request validation error or domain validation error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/LegacyErrorEnvelope" }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Explicit JSON fallback for an unexpected service or registration error. Uncaught framework failures can still use the framework's non-contract response.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ServiceErrorEnvelope" }
          }
        }
      }
    },
    "schemas": {
      "OpaqueId": {
        "type": "string",
        "minLength": 1,
        "description": "Current persistence IDs are opaque strings (typically CUIDs)."
      },
...
      "StringErrorEnvelope": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string" }
        }
      },
      "ZodFlattenedError": {
        "type": "object",
        "required": ["formErrors", "fieldErrors"],
        "properties": {
          "formErrors": {
            "type": "array",
            "items": { "type": "string" }
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": { "type": "string" }
            }
          }
        }
      },
      "ValidationErrorEnvelope": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "$ref": "#/components/schemas/ZodFlattenedError" }
        }
      },
      "ServiceError": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "code": { "type": "string" },
          "message": { "type": "string" }
        }
      },
      "ServiceErrorEnvelope": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "$ref": "#/components/schemas/ServiceError" }
        }
      },
      "LegacyErrorEnvelope": {
        "description": "The v1 routes expose several incompatible error envelopes.",
        "oneOf": [
          { "$ref": "#/components/schemas/StringErrorEnvelope" },
          { "$ref": "#/components/schemas/ValidationErrorEnvelope" },
          { "$ref": "#/components/schemas/ServiceErrorEnvelope" }
        ]
      },
...
      "PublicRegisteredUser": {
        "type": "object",
        "required": ["id", "email", "name", "avatarUrl"],
        "properties": {
          "id": { "$ref": "#/components/schemas/OpaqueId" },
          "email": { "type": "string", "format": "email" },
          "name": { "type": "string" },
          "avatarUrl": { "type": ["string", "null"] }
        }
      },
...
      "RegisterUserRequest": {
        "type": "object",
        "required": ["email", "name", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "name": { "type": "string", "minLength": 1, "maxLength": 100 },
          "password": {
            "type": "string",
            "minLength": 8,
            "x-maxUtf8Bytes": 72,
            "description": "Must occupy at most 72 UTF-8 bytes because v1 uses bcrypt. Character count alone is insufficient."
          }
        }
      },
...
      "RegisterUserResponse": {
        "type": "object",
        "required": ["user"],
        "properties": {
          "user": { "$ref": "#/components/schemas/PublicRegisteredUser" }
        }
      },
...
    }
  }
}