# src/app/api/v1/users/register/route.ts
import { prisma } from "@/lib/db"
import { Prisma } from "@prisma/client"
import { NextResponse } from "next/server"
import bcrypt from "bcryptjs"
import { registrationSchema } from "@/lib/validations/auth"
import { toRegisterUserResponse } from "@/lib/api/v1/response-mappers"

export async function POST(req: Request) {
  const body = await req.json().catch(() => null)
  const parsed = registrationSchema.safeParse(body)
  if (!parsed.success) {
    return NextResponse.json({ error: parsed.error.flatten() }, { status: 422 })
  }

  try {
    const existing = await prisma.user.findUnique({
      where: { email: parsed.data.email },
    })
    if (existing) {
      return NextResponse.json(
        { error: { message: "Пользователь с таким email уже существует" } },
        { status: 409 }
      )
    }

    const passwordHash = await bcrypt.hash(parsed.data.password, 10)
    const user = await prisma.user.create({
      data: { email: parsed.data.email, name: parsed.data.name, passwordHash },
      select: { id: true, email: true, name: true, avatarUrl: true },
    })

    return NextResponse.json(toRegisterUserResponse(user), { status: 201 })
  } catch (e) {
    if (e instanceof Prisma.PrismaClientKnownRequestError && e.code === "P2002") {
      return NextResponse.json(
        { error: { message: "Пользователь с таким email уже существует" } },
        { status: 409 }
      )
    }
    console.error("register failed:", e)
    return NextResponse.json(
      { error: { message: "Ошибка сервера. Проверьте подключение к БД и миграции." } },
      { status: 500 }
    )
  }
}